Refactored connections into StreamApi struct
